pyproject.toml
[project]
name = "patient-api"
version = "0.1.0"
description = ""
authors = [
    {name = "Samuel Whittenberger",email = "sbw79@cornell.edu"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "strands-agents (>=1.6.0,<2.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "opensearch-py (>=3.0.0,<4.0.0)",
    "flask (>=3.1.2,<4.0.0)",
    "gevent (>=25.5.1,<26.0.0)"
]

[tool.poetry]
packages = [{include = "patient_api", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
